Add API version check to plugin loading

Bump the plugin API version and ensure the native loader checks version compatibility before loading.
Also moved library calls into HAL.

Fixes EP-304

~new: Native Plugin Loader now does a version compatibility check before loading native (dll, so) plugins.
~api: Previous plugins are now no longer compatible with platform. These must be updated and recompiled against the latest SDK.
~fix: Calling trim() on an empty string would crash.

See merge request !693

+++ hal/src/driver/qt/components/qmlpluginloader_qt.cpp
@@ -104,7 +104,7 @@ Variant::VarMap QmlPluginLoader::parseTypeDescriptor(QtKernel *pQtKernel, ep::St
         typeDesc.insert("description", ep::SharedString::format("{0} based QML component - {1}", pSuper->asString(), filename));
 
       if (!typeDesc.get("version"))
-        typeDesc.insert("version", (int)EPKERNEL_PLUGINVERSION);
+        typeDesc.insert("version", EPKERNEL_PLUGINVERSION);
 
       return std::move(typeDesc);
     }
